+++ src/main/kotlin/dev/protsenko/securityLinter/docker/checker/DistUpgradeValidator.kt
@@ -2,9 +2,9 @@ package dev.protsenko.securityLinter.docker.checker
 
 import dev.protsenko.securityLinter.docker.checker.core.RunCommandValidator
 
-object AptGetDistUpgradeValidator: RunCommandValidator {
+object DistUpgradeValidator: RunCommandValidator {
     private val pattern = Regex(
-        pattern = """apt-get.*dist-upgrade""",
+        pattern = """\b(apt-get|zypper)\b.*\b(dist-upgrade|dup)\b""",
         options = setOf(RegexOption.IGNORE_CASE, RegexOption.DOT_MATCHES_ALL)
     )