+++ src/main/kotlin/dev/protsenko/securityLinter/docker/inspection/run/impl/PackageManagerAutoYesAnalyzer.kt
@@ -4,15 +4,15 @@ import com.intellij.codeInspection.ProblemHighlightType
 import com.intellij.codeInspection.ProblemsHolder
 import com.intellij.psi.PsiElement
 import dev.protsenko.securityLinter.core.SecurityPluginBundle
-import dev.protsenko.securityLinter.docker.checker.AptGetAutoYesValidator
+import dev.protsenko.securityLinter.docker.checker.PackageManagerAutoYesValidator
 import dev.protsenko.securityLinter.docker.inspection.run.core.DockerfileRunAnalyzerEP
 
-class AptGetInstallWithoutAutoConfirmAnalyzer: DockerfileRunAnalyzerEP {
+class PackageManagerAutoYesAnalyzer: DockerfileRunAnalyzerEP {
     override fun handle(runCommand: String, psiElement: PsiElement, holder: ProblemsHolder) {
-        if (!AptGetAutoYesValidator.isValid(runCommand)){
+        if (!PackageManagerAutoYesValidator.isValid(runCommand)){
             holder.registerProblem(
                 psiElement,
-                SecurityPluginBundle.message("ds019.use-apt-auto-confirm"),
+                SecurityPluginBundle.message("ds019.use-package-manager-auto-confirm"),
                 ProblemHighlightType.WARNING
             )
         }