+++ src/test/kotlin/dev/protsenko/securityLinter/docker/DS004AddInsteadCopyInspectionTest.kt
@@ -2,7 +2,7 @@ package dev.protsenko.securityLinter.docker
 
 import com.intellij.codeInspection.LocalInspectionTool
 import dev.protsenko.securityLinter.core.DockerHighlightingBaseTest
-import dev.protsenko.securityLinter.docker.inspection.add.DS004AddInsteadCopyInspection
+import dev.protsenko.securityLinter.docker.inspection.add.DockerfileAddInspection
 
 class DS004AddInsteadCopyInspectionTest(
     override val dockerRuleFolder: String = "DS004",
@@ -10,5 +10,5 @@ class DS004AddInsteadCopyInspectionTest(
         "Dockerfile-brackets.denied",
         "Dockerfile-empty-add.allowed"
     ),
-    override val targetInspection: LocalInspectionTool = DS004AddInsteadCopyInspection()
+    override val targetInspection: LocalInspectionTool = DockerfileAddInspection()
 ) : DockerHighlightingBaseTest()