+++ src/test/kotlin/dev/protsenko/securityLinter/docker/DS006MultipleEntrypointInspectionTest.kt
@@ -2,10 +2,10 @@ package dev.protsenko.securityLinter.docker
 
 import com.intellij.codeInspection.LocalInspectionTool
 import dev.protsenko.securityLinter.core.DockerHighlightingBaseTest
-import dev.protsenko.securityLinter.docker.inspection.entrypoint.DS006MultipleEntrypointInspection
+import dev.protsenko.securityLinter.docker.inspection.entrypoint.DockerFileEntrypointInspection
 
 class DS006MultipleEntrypointInspectionTest(
     override val dockerRuleFolder: String = "DS006",
     override val customDockerFiles: Set<String> = emptySet<String>(),
-    override val targetInspection: LocalInspectionTool = DS006MultipleEntrypointInspection()
+    override val targetInspection: LocalInspectionTool = DockerFileEntrypointInspection()
 ) : DockerHighlightingBaseTest()